Remove custom SpanProcessor
